edit package
