+++ app/src/main/AndroidManifest.xml
@@ -13,10 +13,10 @@
         android:theme="@style/Theme.MyList"
         tools:targetApi="31">
         <activity
-            android:name=".CategoryActivity"
+            android:name=".ui.CategoryActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
