--- app/src/main/java/com/jrackham/mapper/Mapper.java
@@ -1,8 +1,0 @@
-package com.jrackham.mapper;
-
-import com.jrackham.model.base.Product;
-import com.jrackham.model.realm.ProductRealm;
-
-public interface Mapper {
-    ProductRealm productToProductRealm(Product product);
-}
