--- app/src/main/java/com/jrackham/model/base/Category.java
@@ -1,34 +1,0 @@
-package com.jrackham.model.base;
-
-public class Category {
-    private int id;
-    private String name;
-
-    public Category() {
-    }
-
-    public Category(int id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return "Category{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                '}';
-    }
-}
