+++ app/src/main/java/com/jrackham/persistence/realm/service/CategoryCRUD.java
@@ -1,11 +1,11 @@
-package com.jrackham.persistence;
+package com.jrackham.persistence.realm.service;
 
 import android.os.Build;
 
 import androidx.annotation.RequiresApi;
 
-import com.jrackham.model.realm.CategoryRealm;
-import com.jrackham.model.realm.ProductRealm;
+import com.jrackham.persistence.realm.model.CategoryRealm;
+import com.jrackham.persistence.realm.model.ProductRealm;
 
 import java.util.List;
