+++ app/src/main/java/com/jrackham/ui/ProductAdapter.java
@@ -1,6 +1,6 @@
-package com.jrackham;
+package com.jrackham.ui;
 
-import static com.jrackham.persistence.CategoryCRUD.getCategoryById;
+import static com.jrackham.persistence.realm.service.CategoryCRUD.getCategoryById;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -10,9 +10,9 @@ import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.TextView;
 
-import com.jrackham.model.base.Product;
-import com.jrackham.model.realm.CategoryRealm;
-import com.jrackham.model.realm.ProductRealm;
+import com.jrackham.R;
+import com.jrackham.persistence.realm.model.CategoryRealm;
+import com.jrackham.persistence.realm.model.ProductRealm;
 
 import java.util.List;
