+++ app/src/main/java/com/jrackham/util/mapper/Mapper.java
@@ -1,0 +1,8 @@
+package com.jrackham.util.mapper;
+
+import com.jrackham.model.Product;
+import com.jrackham.persistence.realm.model.ProductRealm;
+
+public interface Mapper {
+    ProductRealm productToProductRealm(Product product);
+}
