+++ app/src/main/res/layout/activity_category.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".CategoryActivity">
+    tools:context=".ui.CategoryActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
